requirements.txt
flask
pandas
scikit-learn
joblib
scapy
speedtest-cli